Vulkan cubemap mipmaps (#22)
* Add cubemap mipmapping

* Add renderer::Fence class

* Refactor

* Materials update textures now

* Use bit mask on octree

* Small changes to make texture blitting work with new changes

* Begin adding PBR to new shaders

* Add normal maps; use staging buffer for mesh

* Add staging buffer pool so we're not creating staging buffers so often

* Add barrier functions and clean up renderer::Image

* Change subpass deps

* Add SSAO

* TMP test case of octree destructor throwing

* Add parallax mapping

* Add atomic counter

* Add SVOs, needs to be cleaned up

* TMP work on SVO

* VCT works, needs multi sampling to make up for mipmaps

* Increase fbo color resolution to use floating points

* Make GraphicsPipeline be Ref<>

* Few changes

* Add svo flag

+++ src/rendering/backend/renderer_helpers.cpp
@@ -6,9 +6,9 @@ namespace hyperion {
 namespace renderer {
 namespace helpers {
 
-size_t MipmapSize(size_t src_size, int lod)
+uint32_t MipmapSize(uint32_t src_size, int lod)
 {
-    return MathUtil::Max(src_size >> lod, size_t(1));
+    return MathUtil::Max(src_size >> lod, 1u);
 }
 
 VkIndexType ToVkIndexType(DatumType datum_type)
